# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "College_Backend",
    "version": "1.0.0",
    "description": "API documentation for College Backend system"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local development server"
    },
    {
      "url": "http://localhost:3001",
      "description": "Alternative local development server"
    }
  ],
  "paths": {
    "/admin/register": {
      "post": {
        "tags": ["Admin"],
        "summary": "Register Master Admin",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string",
                    "example": "admin@gmail.com"
                  },
                  "password": {
                    "type": "string",
                    "example": "123456789"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Admin registered successfully"
          }
        }
      }
    },
...
    "/register": {
      "post": {
        "tags": ["Merchant"],
        "summary": "Register",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "companyName": {
                    "type": "string",
                    "example": "ABC Traders"
                  },
                  "contactName": {
                    "type": "string",
                    "example": "John Doe"
                  },
                  "sellerCategory": {
                    "type": "string",
                    "example": "Electronics"
                  },
                  "contactPhone": {
                    "type": "string",
                    "example": "9710051702"
                  },
                  "contactAltPhone": {
                    "type": "string",
                    "example": "9123456789"
                  },
                  "gstNumber": {
                    "type": "string",
                    "example": "27AABCU9603R1ZV"
                  },
                  "email": {
                    "type": "string",
                    "example": "johndoe@example.com"
                  },
                  "address": {
                    "type": "string",
                    "example": "123, Main Street, Chennai, India"
                  },
                  "password": {
                    "type": "string",
                    "example": "SecurePass@123"
                  },
                  "sellerDocuments": {
                    "type": "string",
                    "example": "document_link.pdf"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Merchant registered successfully"
          }
        }
      }
    },
    "/register/status": {
      "post": {
        "tags": ["Merchant"],
        "summary": "Status of merchant",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string",
                    "example": "karthi582002@gmail.com"
                  },
                  "password": {
                    "type": "string",
                    "example": "123456789"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Merchant status retrieved successfully"
          }
        }
      }
    },
    "/otp": {
      "post": {
        "tags": ["OTP - Verification"],
        "summary": "Send Otp to Phone Using Email",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "contactPhone": {
                    "type": "string",
                    "example": "+919710051702"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OTP sent successfully"
          }
        }
      }
    },
...
    "/buyer/register": {
      "post": {
        "tags": ["Buyer"],
        "summary": "Registration",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "companyName": {
                    "type": "string",
                    "example": "Tech Solutions Ltd."
                  },
                  "contactName": {
                    "type": "string",
                    "example": "John Doe"
                  },
                  "contactEmail": {
                    "type": "string",
                    "example": "johndoe@example.comm"
                  },
                  "password": {
                    "type": "string",
                    "example": "12345678901"
                  },
                  "contactPhone": {
                    "type": "string",
                    "example": "+123456789010"
                  },
                  "billingAddress": {
                    "type": "string",
                    "example": "123 Business St, Cityville, Country"
                  },
                  "shippingAddress": {
                    "type": "string",
                    "example": "456 Warehouse Ave, Cityville, Country"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Buyer registered successfully"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
...
}